In my work over the past year I have encountered a peculiar kind of problem: legacy system which, while not particularly advanced technologically, contained within it a large number of opaque business rules. There was no information provided up front as to why certain logic was applied and it often took trial and error (including errors that affected the users) to discover the origins of this logic. Naturally, whenever a discovery of rule's rationale is made, it should be somehow documented.
In this tutorial we will explore how to provide such a documentation by writing literate, executable specifications using specs2 framework. The focus here is on literacy -- being able to write arbitrary text as our specification and have the engine extract relevant parts of it to run examples and verify results.

We will develop a new component together with its specification from scratch, but similar principles of specification construction apply to legacy systems; in case of the latter some refactoring might be necessary to expose the assembled components of the system in a way that allows us to inject the external dependencies in the tests -- but this is beyond the scope of this tutorial.

We start from a Maven POM that defines initial dependencies: specs2 to define our specifications, JUnit to enable running them from within the IDE, and relevant plug-ins.