Among various ways of structuring the specification, one that is particularly well suited to our scenario is incrementally building the context in which we will exercise the component. This is done by starting with a "given" fragment that sets up the initial context, following up with any number of "when" fragments that transform the context (usually by adding more information to it) and finally a "then" fragment that executes the test and provides the result to the framework.
We start by making the user name and permission table into the initial context. This context is represented by DpcPermission class and the step that constructs has the form of dpcPermissions object. In the specification text, we need to mark the values which should be extracted by the step with '${}' and then insert '^ dpcPermissions ^' to indicate at which point the step should be executed.