Specifying the expected request parameteres and the response of the web service is another step where we just add attributes to the context RequestAndResponse with corresponding "when" step, requestAndResponse.
Note that even though this fragment of the scenario mentions invoking the web service, no action is taken by the spec code -- we'll get down to actually calling the plug-in in the final verification step. We could have invoked the service here and just produced a context containing the response to be verified in "then" step, but by not doing this we are restricting side-effects to the verification step. We'll keep context assembly steps as straightforward as possible so that they can be reused in other specifications.