Clearly, the positive case scenario was insufficient and we have to carry on with the negative one to force the plug-in to do some work. Here we come across the first impendance mismatch between specs2 given/when/then and our approach to writing specifications: "then" fragment ends an example, and to carry on with another one we would have to start again from "given", i.e. set up the context from scratch. That would seriously impact the literary value of our spec, so we'll try a different approach: convert 'accessGranted' to a "then" fragment, but then put an assertion in its body that verifies the condition we were
It is probably possible to solve it in a cleaner way by dropping deeper into the guts of specs2 and creating a custom version of 'RegexStep', but the solution with abusing "when" works, so let's leave yak shaving for later and get back to the task at hand.

We make 'accessGranted' produce a 'DpcPermissions' context -- i.e. forget the document and request/response, but retain the permissions, which should be the same in all examples. After this change specs2 runners no longer report a succesful example, but the spec will fail if the plug-in returns 'false' -- and that's more important.